turns out smart-tab and magit just do not work together now
I kept thinking I'd fixed it because if you do this:

rm -r ~/.config/emacs/elpa
emacs

Then smart-tab works.  If you then close emacs and start it
again *then* smart-tab is broken.  There is also a warning about
smart-tab in Comple-Log on the first "boot" of emacs that indicates
the problem.  Once I figured that out I was able to narrow the problem
down to magit and smart-tab.  So sad.